Move Moo to localtimer package and make VoiceCommand/ExecuteCommandsInBackgroundProcess private
